fix(command): modify the define for the ptr on fct

+++ server/src/util/messages.c
@@ -29,7 +29,7 @@ void send_help_message(int fd)
     char space = '\0';
 
     dprintf(fd, "214 The following commands are recognized.%s", CRLF);
-    for (int i = 0; i < COMMANDS_COUNT; i++) {
+    for (size_t i = 0; i < COMMANDS_COUNT; i++) {
         space = ' ';
         if ((i % 10 == 0 && i != 0) || (i == COMMANDS_COUNT - 1))
             space = '\n';
